test(sveltekit): refactor tests to use new server-common folder
