+++ packages/sveltekit/test/server/rewriteFramesIntegration.test.ts
@@ -2,7 +2,7 @@ import { rewriteFramesIntegration } from '@sentry/browser';
 import { basename } from '@sentry/core';
 import type { Event, StackFrame } from '@sentry/core';
 
-import { rewriteFramesIteratee } from '../../src/server/rewriteFramesIntegration';
+import { rewriteFramesIteratee } from '../../src/server-common/rewriteFramesIntegration';
 import type { GlobalWithSentryValues } from '../../src/vite/injectGlobalValues';
 
 describe('rewriteFramesIteratee', () => {
